ta_autograder/requirements.txt
openai>=1.30.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
